+++ CHANGELOG.md
@@ -11,6 +11,12 @@ The [releases page](https://github.com/Jelmerro/Vieb/releases) also contains the
 but the list below contains much more technical details.
 The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
+## Unreleased
+
+### Removed
+
+- Redundant ozone-platform-hint=auto flag from builds now that it has become a Chromium default
+
 ## [12.5.0](https://github.com/Jelmerro/Vieb/compare/12.4.0...12.5.0) - 2025-09-03
 
 [code diff](https://github.com/Jelmerro/Vieb/compare/12.4.0...12.5.0) - [released builds](https://github.com/Jelmerro/Vieb/releases/tag/12.5.0)